requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==4.0.0
asyncpg==0.28.0
certifi==2023.7.22
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iso8601==1.1.0
psycopg2-binary==2.9.7
pyaes==1.6.1
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
pypika-tortoise==0.1.6
python-telegram-bot==20.5
pytz==2023.3.post1
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.21
Telethon==1.30.3
tortoise-orm==0.20.0
typing_extensions==4.7.1
